ml_model/requirements.txt
ultralytics
Pillow
python-dotenv
numpy
opencv-python
dvc[s3]
labelbox
labelbox[data]
ndjson
scaleapi
albumentations
torchmetrics